Apply environment natively in addition to JVM view - fixes #3
The JVM has it's own special view of environment variables, which was
the only thing changed previously. That does not work, unfortunately, if
any child processes are spawned, which can happen in SBT if new JVMs are
started by forking, or if artifacts are package before execution.

To change the native environment we rely on JNA, with two different
implementations, one for Windows and one for POSIX platforms, which
should be good enough coverage. We call the corresponding C library
functions to alter process environment.

Unfortunately, we need to keep the old hack as well since the JVM never
updates its view of the environment after startup.